+++ app/models/scenarios/scenario_417.rb
@@ -55,6 +55,7 @@ module Scenarios
           map_data:,
           allied_units:,
           axis_units:,
+          special_rules: ["axis_bonzai"],
         }
       end