+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Allow infantry fire into unarmored side of vehicle
 - Compress stacks too big for display
+- Bonzai rules for 417
 - Updated/promoted 315
 - Promoted 010, 307, 504 to beta
 - Added prototype 419, 424